Execute model query for the new decorator constraint that matches against the resolved callees
Summary:
The main changes are in function `matches_decorator_constraint` of
`modelQueryExecution.ml`. We can now match against the newly added constraint
from the previous diff. The implementation reuses function
`resolve_callees_from_type` from `callGraph.ml`, to resolve callees based on
types of the callee expression.

At this moment, the override graph is set to an empty graph when calling
function `resolve_callees_from_type`, which means it wouldn't be able to
precisely resolve calls that require an override graph. But this should be fine
for the use case of resolving calls to decorators, since it is probably a rare
case that a decorator has overrides.

Reviewed By: arthaud

Differential Revision: D54130363

fbshipit-source-id: d87e7ae3d0511f7086f14653c43c58f8d00ce41e