Don't use `all_module_paths` for cache invalidation
Summary:
It does not make sense that Pysa is using `all_module_paths` for cache
handling; the difference between `all_module_paths` and `module_paths`
is that the former includes shadowed module paths, but those are not
relevant to Pysa because they cannot be analyzed.

From digging around years-old commit messages I think the reason this happened
is that the logic used to be part of the saved state code for Pyre and
was dropped into Pysa as-is.

Reviewed By: grievejia

Differential Revision: D54258787

fbshipit-source-id: b172e73d36d6e1459817abd7b7cc4f7ed274e5ab