Use pyre pysa api in fetch-callables logic
Summary:
Push the api down, and convert most api calls to use it rather
than grabbing Pyre environment layers directly.

We're still extracting a GlobalResolution.t inside `gather_raw_definitions`
because I haven't yet converted the `Target` module; I'll have to do that
close to the very end of the stack because `Target` is used in so many
places; we'll have to rely on extracting a GlobalResolution.t until then.

Reviewed By: grievejia

Differential Revision: D54289461

fbshipit-source-id: 2428c91c7fd5fb5035b6056ca1bd001f4456e39b