Flatten save_results_to logic
Summary:
The Pysa `type_check` function control flow is more complex
than it needs to be because it has two passes of gathering
data (qualifiers, then define names) and dumping the data to
disk conditionally.

The ordering of these operations isn't important, so we can simplify
the control flow by gathering both bits of data and only using
a single `if` branch to write both to disk.

Reviewed By: connernilsen

Differential Revision: D54213130

fbshipit-source-id: c92f760cc8914737e1b550556f72c04c9afe73d5