Begin work on the ReadOnly pyre/pysa api
Summary:
**What this diff does**

This commit bootstraps an initial read-only pyre/pysa api.

At this stage it doesn't do anything (we just extract
the global module paths and read-only type environment), but
it ensures that we've covered all the read-write interfaces.

**Adding `purge` to the read-write API**

In the process I caught one read-write operation that escaped
the previous diff, which is purging shared memory to drop
raw Sources; this is something Pysa does because the raw
sources are fairly big, and are never used for taint analysis
(Pyre only keeps them because we need them for post-processing
analysis errors, Pysa doesn't need this).

**Next steps**

I'll stack diffs on top pushing this new api down into business logic
and gradually migrating calls to a "flat" API instead of extracting
Pyre environment layers to query them directly.

Reviewed By: grievejia

Differential Revision: D54271560

fbshipit-source-id: b1d801539a7c4b3246fea5feeb9de12e39432d23