Move source path conversion code
Summary:
Up until now all source-path related code has been in `Server.PathLookup`.

I think the underlying motivation is that in principle analysis code never
needs to work with a SourcePath.t, and so it doesn't belong in Analysis;
this is a fair critique in the absence of other concerns.

But keeping it in `Server` on the basis of ^^ is clearly worse; we're
leaking those functions into both Pysa code and `code_navigation_server`,
which is a strong indication that they are more fundamental operations
than belong in the legacy `server` codebase.

Also, I think the big argument against keeping them out of `Analysis`
wasn't hygiene but safety; at one point in the past both source and
artifact paths were just `PyrePath.t`, which made it very hard to
avoid mixing them up, so we used module structure to try to enforce
safety. These days `SourcePath.t` and `ArtifactPath.t` are distinct,
so it's no less safe to have these conversions in a more accessible
place.

My immediate motivation for making this change is that I don't
want Pysa code to have direct access to a `SourceCodeApi.t`, but
if these functions remain inside the `Server` module I won't
actually be able to extract them - which is yet another indication
on top of the preexisting exports suggesting that `Server` is not
a great place for this logic.

Reviewed By: grievejia

Differential Revision: D54264728

fbshipit-source-id: 746a0887f4c7a69d23d3fa0e791b6032602ef794