Split pysa tests
Summary:
In the middle of my refactor to use PyrePysaApi I hit an unexpected
behavior change (in retrospect, the issue is that global queries in UGE depend on us having pre-populated the cache tables which I wasn't doing in my new `Test.ScratchProject` function; it was easily fixed once I could see what was happening).

But I couldn't actually debug because, as usual, the flat tests make it so
that failures are pretty challenging to isolate and the logs are way too big
if I try to print debug.

So, doing the usual labeled test case refactor.

Reviewed By: grievejia

Differential Revision: D54279487

fbshipit-source-id: 98c15c918ef0136834169db718167c9364fb6def