Create initial PyrePysaApi with ReadWrite module
Summary:
This diff kicks off a stack where I pull the Pyre <-> Pysa
boundary (or a boundary; the lines are blurry and we have to
decide where the boundary should be so part of this stack is deciding)
into a single module.

There are two goals of doing this:
- In the short run, it will make refactors of the Pyre environment easier
  because, as long as the refactor doesn't change actual behavior Pysa
  depends on, we'll usually have a single module to update for Pysa instead
  of having to make changes throughout `interprocedural` and
  `interprocedural_analysis`
- In the longer run, we need to plan for how Pyre and Pysa are going to
  interact, particularly if Pyre makes major changes such as reworking
  how symbol resolution works to eliminate qualification or changing
  our representation of Python types. Having this flat API will give us
  a starting point for documenting all of Pysa's current uses of Pyre,
  and thinking about how to make changes.

The new `PyrePysaApi` module is going to contain at least two, maybe
three sub-modules:
- It needs a module representing how Pysa runs a type check and handles
  loading from a cache. This is the `PyrePysaApi.ReadWrite` module
  defined in this diff.
- It also needs at least one read-only api representing all the points
  at which core taint analysis access the type environment. I don't
  yet have a diff with this.
- We may or may not want a second read-only API with the (different,
  typically lower-level) interfaces needed in model queries; I'll
  decide later whether we want two read-only APIs or just one.

This diff creates only the `ReadWrite` api. It decouples `Cache.ml`
entirely from `TypeEnvironment` and lower-level Pyre environment layers,
but leaves downstream logic in `taintAnalysis` alone, since that logic
needs to migrate onto a new read-only interface that does not yet
exist.

I'm deferring making the `.mli` file for a little bit to reduce churn as I work
through the stack; I'll definitely add one later.

Reviewed By: alexkassil

Differential Revision: D54264690

fbshipit-source-id: 5ef43df7618b6075aab92caaad001e2c097158bf