Push ModuleTracker access out of ChangedPaths
Summary:
It's easier to refactor how Pysa accesses Pyre information if most
of the cache related accesses are pulled up to the top level
`taintAnalysis` and `Cache` modules. It's pretty easy to make
`ChangedPaths` a module of functions taking the module paths
as input rather than directly accessing module trackers, so
I'm making that change.

Reviewed By: grievejia

Differential Revision: D54259561

fbshipit-source-id: 09b874b7c7bb3eb2cbf288a5128b23ba36019041