Create a contextual api, use it in traversals
Summary:
This diff is probably the "meat" of the refactor to thread all
Pysa use of Pyre through one layer; everything that came before
was getting to the point where we could create and use a
`Resolution.t` - like abstraction for Pysa, and everything that
remains will follow from continuing to push this new abstraction
down until there are no remaining direct Pyre environment layer
calls.

It's a big change, but also pretty mechanical and actually requires zero
changes to tests, which is a pretty good sign that it is correct.

**Explanation of the new datatype `PyrePysaApi.InContext.t`**

I've reached the end of what I can do in refactoring Pysa
to use a flat read-only Pyre api; the remaining code changes
are going to involve at least sometimes accessing Pyre information
in a scoped context (that is, at a specific point in an ast traversal,
which we do by way of using the statement keys in Fixpoint logic)

To support these use cases, I need a `Resolution.t` - like construct,
but specialized for Pysa (I don't want Pysa directly using `Resolution.t`,
because you can extract a `GlobalResolution.t` from that which means
we're not actually routing Pysa calls through one place which is the
whole point of this refactor).

The new type, `PyrePysaApi.InContext.t`, wraps both a global
`PyrePysaApi.ReadOnly.t` and a `Resolution.t` so that it can provide
scope-specific information.

Much like `Resolution.t`, we need two key capabilities here:
- First, we need to be able to create an in-context Pyre handle
  keyed on a particular statement in the fixpoint; this is basically
  what we have instead of a typed AST, so this is mandatory.
- Second, we need some monadic apis for traversing expressions that
  can require new context; this is because given the lack of a typed
  ast and the substitution of statement-keyed "resolutions", we have
  to be able to reconstruct any side-effects from traversing expressions
  (By the way I *hate* that Pyre does this, but it won't be easy to
  fix so we have to live with it for now). At the moment Pysa only
  actually deals with traversing implicit assignments from generator
  expressions, so that's the only call we need thus far.

**What does this diff actually do**

We create `PyrePysaApi.InContext.t` and we thread it through all of the
traversals in the forward and backward fixpoint.

The `LineageAnalysis.forward_analyze_call` function is mutually recursive
with forward analysis (by way of a callback so as to not have circular
dependencies, but nontheless it is mutually recursive) so we have to
update it in the same pass.

We do *not* yet modify any other lower-level code; whenever we hit the
boundary we extract the underlying resolution; later diffs will push
the new abstraction down and add more interfaces as needed.

Reviewed By: arthaud

Differential Revision: D54335096

fbshipit-source-id: e249b56cb5e87c66df55b3333ca2ef782a151eef