Fix upgrade add_local_mode when file starts with an ignore comment
Summary:
In `filesystem.add_local_mode()`, we would add any local mode comments after any 'headers' when it was appropriate for one to be added. Headers were considered to be any comments before non-comment or whitespace lines that weren't `# pyre-ignore`, `# pyre-fixme`, or `# type: ignore`, with any amount of whitespace surrounding it.

The issue with this is that the suppressions we land are typically of the form `# pyre-fixme[X]`, which does not match the bare `# pyre-fixme` we're matching on. That results in mode header comments being added after suppression comments if the suppression comments are on the first few lines of a file. We'll also then be splitting up a suppression comment and the line it's suppressing, leading to two new warnings for an unused suppression and the error that was being suppressed.

To fix this, we now need to do a match-all wildcard after the suppression, but make sure we're not including the similar `pyre-ignore-all-errors` comment in that.

Reviewed By: grievejia

Differential Revision: D54389289

fbshipit-source-id: 5c5c590c40c04680bdd3bc17af7b0d5a6c797928