Factor out get_server_start_command function
Summary:
This diff factors out the logic deciding on the command to start the pyre server. It is shared between start (the entry point to start the server via terminal commands) and initialization (the entry point to start the server via code-navigation / persistent). Ideally, this logic will be merged. but that is not in the scope of this stack.

This stack is aiming to add a cgroup for each pyre daemon process similar to [flow's](https://fburl.com/code/4sms9xjn) in order to mitigate the negative effects of S395002.

Reviewed By: inseokhwang

Differential Revision: D54364242

fbshipit-source-id: aaa8b80d026e9501bb91194db2826c09b18c0158