Route dangling Analysis queries through PyrePysaApi
Summary:
We don't want pysa code directly using `GlobalResolution.t` or `Resolution.t`
because we want a single module (`PyrePysaApi`) to capture all of the
environment access, which will help us with refactoring and with planning
bigger architecture changes in the long run.

So, let's proxy these calls through PyrePysaApi.

Reviewed By: grievejia

Differential Revision: D54372914

fbshipit-source-id: f368a7596a027c950ba1f0475d8c9571d92076e1