Tweak the pyre-in-context-at-key constructor
Summary:
In addition to ForwardAnalysis and BackwardAnalysis, we need
a context-awere Pyre handle in callGraph as well.

But in that context we have a `Statement.Define.t`, not a
`Statement.Define.t Node.t`. So, pushing up the extraction of
the node value so that our constructor will be usable in both
cases.

Reviewed By: grievejia

Differential Revision: D54361857

fbshipit-source-id: fa902dcca3a89359276fd32bfc7fde6bd691f824