Include files dropped from sourcedb in classic.bxl query
Summary:
In doing one last check before rolling this out, I noticed that some files weren't getting included when running the classic.bxl script for getting files. On looking into it, I realized the script wasn't prepared for getting any files that are dropped by the sourcedb.

This makes sure to add in any files that get dropped, so we should be migrating all files in a project.

Reviewed By: grievejia

Differential Revision: D54422747

fbshipit-source-id: 0973aac4371579467c0d6294c4312423b174a0b4