Fix off by one maximum trace length
Summary:
The way we limit the trace length is by filtering taint with a distance greater or equal than the maximum trace length in models.
However, we should actually limit by `maximum_trace_length - 1`, since the distance will be incremented by one when propagated.
With the current implementation, we are seeing issues with trace length 2 when the maximum trace length is 1 (see integration test).

Reviewed By: tianhan0

Differential Revision: D54752661

fbshipit-source-id: 02aa60c104d825f93086a6a5f38cf13d407a3028