Initiate a visitor which will return a list of document symbols
Summary:
Context:
We are supporting document symbol in the backend.

Changes:
Set up all methods needed for getting document symbols from an overlay and translating the response to a type which corresponds to what's in response.ml

Set up a method with access to the source file in order to walk through the AST

In the method which processes the AST, I printed the source file into the log. I know this will crowd the log, so I intend to remove it immedietly in the diff to follow, which will contain the actual implementation. We don't have to worry about crowding the logs with this, since I did not enable the call to the backend.

Reviewed By: kinto0

Differential Revision: D54550470

fbshipit-source-id: 0428754fd9c1e9bb2ecf1c49ac5c6f171bbab219