Fix ModuleNotTracked errors
Summary:
# Background

In our logging, Kyle noticed that one of the most common errors is `ModuleNotTracked`, which appears to only happen on Global Lazy servers. Regular codenav does not appear to have the issue.

After looking more into it, this is what I found:
1. `ModuleNotTracked` happens when the language server requests information on a file that doesn't exist.
2. When Global Lazy requests type errors, we send requests for all currently opened files.
3. VSCode lets us know when a file is deleted *if* its deleted from within VSCode, but some users delete files from the command line or in other ways.

# Problem

Therefore, we're running into this issue where a user deletes a file outside of VSCode, but keeps the pane for that file opened even though the file no longer exists. Since we never receive a file closed or file deleted notification, we still assume the file exists, and request information on it.

# Solutions

## Solution A

This diff attempts to fix the `ModuleNotTracked` errors that we're getting by ensuring that, for codenav-based type errors **only**, we check that the file exists before including it in the request.

## Solution B

There's an alternative solution we can take that would be cleaner, but would require a lot more setup. We can create a VSCode `FileSystemWatcher` to notify us of any delete operations that happen to Python files outside of VSCode (although this may also include deletes that happen within VSCode as well).

```
const fsWatcher = workspace.createFileSystemWatcher('**/*.py', true, true, false);
fsWatcher.onDidDelete(uri => {
   logger.info(`File deleted :) : ${uri.path}`);
});
context.subscriptions.push(fsWatcher);
```

The issue with this is that we'd need to add extra request/response infrastructure for handling these notifications on the language server and language client sides, but I don't think it's worth the effort to do it if this is something that only happens in rare cases and only when running global lazy.

Reviewed By: grievejia

Differential Revision: D55204102

fbshipit-source-id: 1c666524fa52502fb23c733499118f6c5682144a