Restart Pyre language server before processing request if needed
Summary:
One of the most common reasons for a `ClientNotRegistered` error seems to happen when an exception occurrs in the daemon manager that causes it to unregister the language server from the daemon.

On the next code navigation request, the language server will query the daemon and find out that it got unregistered by the daemon manager, then skip processing the request.

In this diff, we'll try to restart the daemon manager **before** each request as well as after each request (for the cases we were already attempting to restart the daemon.

Reviewed By: grievejia

Differential Revision: D55380232

fbshipit-source-id: 9fba037867d508d86bacd624887ae8afc65ab6a9