Add additional error types for ocaml failures added in D55383923
Summary:
## This Stack
Adding additional error logging so we can better breakdown causes of UPND

## This Diff
Additional error logging was added in D55383923, this diff takes the new errors that gets logged, maps it to `PyreLanguageServerError` which will get sent to our telemetry logging.

## Next Diff
Update this so that rather than doing a substring match, we are able to accept an error type from daemon and use it directly.

Reviewed By: samwgoldman

Differential Revision: D55461385

fbshipit-source-id: 1b00ee32c87c979bce233264a6a6c67d2c01f0ec