Add more locations for setting and clearing client_register_event
Summary:
I think the event not being reset is the cause for why ClientNotRegistered errors occur so much more in global lazy. The pattern that happens is probably this:
1. the language server, daemon manager, and daemon start up, the language server gets registered, and the asyncio.Event that stops requests from being sent before registering the language server is set
2. the daemon and/or daemon manager die, but the language server stays up, the asyncio.Event is not cleared
3. any requests the user makes are queued up and sent to the daemon, these requests are not blocked by the asyncio.Event since it was already set
4. the language server begins to start up the daemon manager and daemon
5. the daemon gets a bunch of requests queued by the language server before it gets registered, resulting in ClientNotRegistered
6. the language server gets registered

I originally added the asyncio.Event to help with the ClientNotRegistered error, since originally that was happening way more with type error requests when global lazy was first starting out. I think the issue here is mostly that I didn't realize the restart logic could keep the language server alive in some cases, which would result in issues if the asyncio.Event isn't cleared. 😢

I think some of the other diffs that have been added will also help reduce the ClientNotRegistered, but fixing this (which should take another hour or two to redo a part of it) will be the biggest improvement.

Reviewed By: grievejia

Differential Revision: D55388577

fbshipit-source-id: 024a3d73c359b3fa995f84990b902dd1683cd4c5