Stop RemoteBackedIndexQuerier from disposing its base_client
Summary:
In the codenav server, when running [`client_teardown()`](https://www.internalfb.com/code/fbsource/[be367beb9f0d8c68d33fc21b95950b4900a31b0c]/fbcode/tools/pyre/client/commands/code_navigation.py?lines=176-181), we call `handle_dispose_client()` for each querier active.

In `RemoteBackedIndexQuerier`, `handle_dispose_client()` just calls `handle_dispose_client()` on its `base_querier`, which, in the codenav server, is a `CodeNavigationDaemonQuerier`.

The same `CodeNavigationDaemonQuerier` is also one of the queriers being disposed of, so in `client_teardown()`, it ends up getting disposed of **twice**. This results in the second `DisposeClient` request in the daemon returning a `ClientNotRegistered` error, which has been affecting the codenav server metrics.

I'm not sure if this is the approach that we want to be taking here (if we're trying to leave this abstract), but a better approach would likely require a larger refactor.

Reviewed By: dluo

Differential Revision: D55430712

fbshipit-source-id: aa87e5d53925759422e3bdd3292270b5490326a4