Update locationBasedLookup to contain logic for checking is a location contains a position
Summary:
Context
The goal is to fix a bug in GoToDef where if the carat is at the end of a valid definition, GoToDef still fails as shown here:
{F1476321497}

We identified that the bug is in the "contains" method in the location class. The fix contains two parts

1- Have the logic for checking if a position is contained in a location in locationBasedLookup.ml and use it in all LSP calls

2- Edit the method to addrss the bug

Changes:

Have an outer level function location_contains_position and update the callers used for the LSP to use that function in locationBasedLookup.ml

We will follow up by modifying that function to address the bug in the next diff.

Reviewed By: samwgoldman

Differential Revision: D55608473

fbshipit-source-id: f7b40d3ab4b283343ef66a4a5affbbd53d2b75eb