Change the function which checks if a location is contained in a position inclusive
Summary:
Context
The goal is to fix a bug in GoToDef where if the carat is at the end of a valid definition, GoToDef still fails as shown here:
{F1476321497}
We identified that the bug is in the "contains" method in the location class. The fix contains two parts
1- Have the logic for checking if a position is contained in a location in locationBasedLookup.ml and use it in all LSP calls
2- Edit the method to addrss the bug

Changes:
1- We create a function location_contains_position_inclusive, which is similar to location_contains_position. As an example, the effect of it is that the range 1:0-1:2 includes the position 1:2.

We use these functions in two places within find_narrowest_spanning_symbol which is when visiting an argument or an expression. We leave the remaining occurrences unchanged till future discussions.

2- We abstract over location_contains_position_inclusive and location_contains_position

3- We address a bug in autocomplete where we change the column number of the autocomplete request, which is the wrong behavior

4- We disable autocomplete tests till we fix the feature

5- We reflect the new behavior in existing unit tests

Reviewed By: kinto0

Differential Revision: D55611309

fbshipit-source-id: 60913cd8dd773a1d35651752f010c7cfe3453ea5