Simplify the initial comprehension test case
Summary:
As in the previous diff, I'm trying to make the unit tests more suitable
for use in understanding what Pyre is doing under the hood. The previous did
this by isolating test cases, a prereq for print debugging, and this diff
further simplifies by using a more minimal example with just one statement
to traverse.

Nothing relevant to the test case is lost.

Reviewed By: samwgoldman

Differential Revision: D55706927

fbshipit-source-id: c43555d570eba5e563a66cd1887bf855a8535f1a