Refactor test_lookup_comprehension
Summary:
I'm performing the same conversion to labeled test cases that
I always do now before starting on backend work; in this case I'm
isolating the (passing) tests verifying that we correctly resolve
type information in batch mode for comprehensions.

I want these tests to be easy to run so I can use print debugging to
understand how it works, which is not possible when we run dozens
of tests per test case (the logs become spam).

Reviewed By: samwgoldman

Differential Revision: D55706013

fbshipit-source-id: 7d2c1adfe751980acfdeef7dcd7dd257dc19f0b3