Add parameter source modeling capabilities
Summary:
Previously we could only model generation sources on captured variables models.

Modeling a captured variable as a source means one is modeling this:

Note: we cannot infer a generation source `captured_variable(...)` without having an equivalent parameter source somewhere in the model.

```
def outer():
  x = ...
  def inner(): # inferred model is "source: captured_variable(outer.x)"
    nonlocal x
    x = source()
    sink(x) # place 1
  inner()
  sink(x) # place 2
```
and in this situation, we infer issues in both place 1 and place 2.

So now to support the equivalent for `CapturedVariable(TaintSource[...])` let's have it act as both a parameter source and generation source, like it does for the inferred case.
```
def outer():
  x = ...
  def inner(): # user defined model is "source: captured_variable(outer.x)"
    sink(x) # place 1 <- this diff, aka parameter source
  inner()
  sink(x) # place 2 <- already supported, aka generation source
```
vs
```
def outer():
  x = ...
  def inner(): # no model
    sink(x) # no issue
  inner()
  sink(x) # no issue
```

Reviewed By: tianhan0

Differential Revision: D55660448

fbshipit-source-id: e333e64cf4a59dd8118af4d953763290d7d88b8d