Demonstrate a false negative due to resolving calls to `str` on optional types
Summary:
When we see a call to `str(a)`, we have special logic to decide if that will call `a.__str__()` or `a.__repr__()`.
That logic does not handle the case where a has an optional type such as `Optional[Foo]`.
It will wrongly resolve to a call to `a.__repr__()` even if `Foo.__str__` exists.

This test demonstrates the problem. We will fix the issue in a follow-up.

Reviewed By: r0rshark

Differential Revision: D55738773

fbshipit-source-id: 72809b31cb2fb5719eaf1abf2fbf3e8f3a8ca426