Return empty_reason with definition response
Summary:
Instead of returning an error response, return a definition response with an
associated empty reason.

Before this diff, we sent the empty reason as an error response, which caused
the client to perform a fallback query to a remote index. This change restores
the previous behavior where empty responses had no fallback.

Reviewed By: kinto0

Differential Revision: D55731548

fbshipit-source-id: 7da2f72ea447cbbbf21d54a0c798fa42d9b3c026