Add empty_reason to LocationOfDefinitionResponse, log with telemetry event
Summary:
This diff adds a new, optional field to definition responses, so the backend can
provide a reason for an empty response. The field has a default value and will
currently always be None because the backend doesn't return anything yet.

I changed the dataclass_json config to use snake_case instead of camlCase,
because it's idiomatic to use snake_case in OCaml. That makes this work with the
next diff where I add the response.

Reviewed By: kinto0

Differential Revision: D55731547

fbshipit-source-id: 2c7f934303821c4b73edecb5fb19ea79e84f782f