Use the resolve definition from position logic in tests
Summary:
I'm not sure why - maybe because the tests were written prior to the end-to-end
function - but the unit tests were not actually using the production code that
produces a definition location from a position. Instead, they were calling the
lower-level helpers directly.

This is confusing and makes the tests harder to skim, so cutting them over to using
`location_of_definition`.

Reviewed By: grievejia

Differential Revision: D55746829

fbshipit-source-id: 095833904446f8077607b919cf724cbba7994fcf