Add glean and daemon query subtimers
Summary:
We already track query_duration in our telemetry event. This diff adds two
subtimers: daemon_duration and glean_duration. This will make it easier to
attribute latency.

I needed to pass durations through the stack, which required a bit of additional
wiring. The glean interface is abstracted behind a "remote index" interface, so
I extended the interface to return a duration.

I also added a duration to the DaemonQueryFailure object. I somewhat lazily gave
this a default value to avoid updating callers everywhere. This could be
improved, but I'm leaving that for later work.

Reviewed By: migeed-z

Differential Revision: D55783975

fbshipit-source-id: 6b4b3ed6225128e4f99b193ad49b96cf970e81b7