Log durations from daemon in telemetry event
Summary:
We track durations for how long the the client waited for a round-trip to the
daemon. This diff adds the duration that the daemon spent answering the query
once it started.

This logging will help diagnose where we are spending time. If the daemon
"inner" duration is significantly less than the daemon duration observed by the
client, then we should look at the architecture of the daemon to ensure it's
responsive to requests.

Adding a non-deterministic value to responses adds a complication for automated
testing. In this diff, I add a "scrubber" which changes all measured durations
to 0. To build this scrubber, I needed to add Response.of_yojson by updating the
various PPX from "to_yojson" to "yojson".

Reviewed By: kinto0

Differential Revision: D55788169

fbshipit-source-id: 09bc82d4c35d69e7dcbc5b041bba2ceffda5c0d8