Parse maximum_source_distance and maximum_sink_distance options in rules
Summary:
We want to be able to filter out issues before they are emitted to SAPP, for some of our most noisy rules.
This will allow us to reduce the number of issues we emit and store in our database, and therefore reduce our database size.

This diff stack adds two options `maximum_source_distance` and `maximum_sink_distance` in rules.
This is not meant to be a full blown filtering system, for that users should still use SAPP filters.

This specific diff implements the parsing of those options in rules.

Reviewed By: tianhan0

Differential Revision: D55747069

fbshipit-source-id: 1862febd80d046e1d2b379cc84512df463d2e0a7