Skip migrating files if they're covered by a nested local config
Summary:
# Background
An issue with the previous version of the configurationless migration script is that it expects *exactly* one .pyre_configuration.local to cover each file.

Due to how some projects are structured, this isn't the case. There can be nested .pyre_configuration.local files, where the parent configurations will apply a `# pyre-ignore-all-errors` suppression to files covered by a nested configuration.

In the initial migration, the parent configurations were applied first, so we need to add a way to only apply the closest configuration's mode and force the re-application of the Pyre mode.

# Approach
The approach I took to solve this is to use a glob pattern to find nested configurations, then use the same process for finding the files to migrate to find files that should be covered by nested configurations. Then I filter out any files that will be covered by a nested config.

Reviewed By: stroxler

Differential Revision: D55651659

fbshipit-source-id: dd2aa00d4fadffb7548b05c07b76bf9d59b42242