Remove useless short circuit check
Summary:
Back in an early version of the configurationless script, I added a short circuit to get out of processing if we ever reached a local or global configuration with no ignored/excluded files that matched in the strictness mode they ran in by default. This was done before we decided that it would be better to make the mode header required in files that are being type checked so that the CI/CLI type checkers and IDE type checker worked mostly in sync.

Since we decided to make mode headers required on type checked files, that assumption is no longer true. We always need to migrate all files under a configuration unless every file has already been migrated, which we can't know without inspecting all of the files.

Luckily, since `exclude_patterns` and `ignore_all_errors_prefixes` are never empty (internally), we've always been skipping this short circuit. To clean things up and make sure we don't accidentally cause issues for ourselves in the future, let's get rid of the short circuit entirely.

Reviewed By: grievejia

Differential Revision: D55878124

fbshipit-source-id: 26a8cce495bd2153ec6d0238debdf7812186cc78