Fix dataclass transform bug [1] use correct symbol table
Summary:
This commit fixes the bug illustrated in the previous commit: Pyre
wasn't able to resolve dataclass transform decorators defined outside
of the project because it was using the wrong lookup table; we ought
to be using the `UnnanotatedAttribute.t` table rather than the
`FunctionDefinition.t` table to check decorators because only the former
is populated on dependencies that don't need type checking.

Note that we are now able to omit get_function_definition from the signature
of AttributeResolution.Queries, which is good because it would *always* be
a bug to use that function (which will only return a body if the function
is in our project / scheduled for type checking) in attribute resolution
logic (which should always work correctly on dependencies that aren't part of
the project).

Note the extra test case verifying that if a dataclass transform has
overloads this works; I don't think overloads are likely on a dataclass
transform but it's good to ensure that we are going to use the
correct decorator list (the one on the implementation) in that case.

Reviewed By: grievejia

Differential Revision: D55900528

fbshipit-source-id: 545f298d96467c96eaf799deb67744cfad1d2305