Fix dataclass transform bug [0] extend test cases
Summary:
As he was looking at type check results when we more agressively
break up projects, Jia noticed a strange pattern: changing the
project-vs-dependency boundary modifies dataclass transform behavior!

It turns out this is due to a misuse of our symbol tables in the
decorator-resolving logic; attribute resolution should never be
using the `FunctionDefinition.t` values from UnannotatedGlobalEnvironment
because those aren't actually populated unless a source is scheduled
for type-checking (this is a hack that saves us quite a bit of RAM,
since the function bodies are fairly large).

This commit is the first half of a fix: create a new test case illustrating
the bug. The fix will be stacked.

Reviewed By: grievejia

Differential Revision: D55900529

fbshipit-source-id: f893c2ab4c2326ee1a0295b9e4afcf688d92954b