Rename some UGE functions
Summary:
We just fixed a bug handling `__dataclass_transform__` whose root cause was the
misuse of `get_function_definition` / `get_define_body`, which shouldn't be
called except in type checking or Pysa because it doesn't work on sources that aren't
considered "in the current type checking project" (for Pysa, everything is).

This commit makes the situation less risky by:
- renaming these functions in Pyre to explicitly indicate they only work "in project";
  I don't think the names are self-explanatory per se but I do think they will now
  likely cause someone to double-check what these do instead of assuming they just work.
- adding comments on both the implementations that go-to-definition would jump to and
  the interface files of UGE and globalResolution indicating what these functions do
  in plain English
- including comments in PyrePysaApi indicating why it's safe to use these functions
  and why we can drop the `_in_project` part of the name.

In addition I moved `get_define_body` into globalResolution; it is only used from there
(never directly on UGE), and in general trivially derivable functions should live in
globalResolution when possible to avoid cluttering the environment code, which is usually
trickier to read / understand / maintain.

Reviewed By: grievejia

Differential Revision: D55923841

fbshipit-source-id: f8a89fcdbc787b09a8352b1bec8f5807c05634b9