Move Pysa-only `source_is_unit_test` out of `globalResolution.ml`
Summary:
This is only used by pysa. One benefit of stroxler 's work refactoring and adding the pyre-pysa-api is now it is easier to see and split these away.

I considered having this implementation live in `pyrePysaApi.ml` or in a new file `pysaApi.ml` or `pysaResolution.ml` or something. I decided it having it just part of `pyrePysaApi.ml` is good enough.

Reviewed By: arthaud

Differential Revision: D55905919

fbshipit-source-id: 142d3c3d67a69db87a78f56ae72e27d70df5b9bc