Remove unused flag `include_unit_tests` for FetchCallables
Summary:
When reading this code, I noticed this flag was only true in one unit test, and changing the flag had no changes in behavior.

All this does is overwrite the behavior of skipping this check
https://www.internalfb.com/code/fbsource/[5fd0d001a3aa]/fbcode/tools/pyre/source/analysis/globalResolution.ml?lines=503-544
which ignores analyzing unittests and pytests.

We didn't analyze those previously, so no need for the special flag. If we start caring about this for Pysa, we can just remove the call to `PyrePysaApi`

Reviewed By: arthaud

Differential Revision: D55904340

fbshipit-source-id: 442a6bcc7a9abffbbf2d1a16abdd74f56d83112f