Make it much clearer how dataclass transforms work [refactor + bugfix]
Summary:
I would like to improve our dataclass transform support - at least handling the
normal names `typing.dataclass_transform` and
`typing_extensions.dataclass_transform` so that we can more reasonably see what
other gaps remain.

But when I started working on this, I realized that the complexity is nontrivial:
- the call to a `dataclass_transform` is either a decorator-of-a-decorator or
  a decorator-of-a-base-class
- so we're always processing both that original decorator at the custom transform declaration
  *and* some other code at the place where a custom transform is *used*
- this means that poorly factored code, or even well-factored code without clear human-readable
  comments, will be pretty hard to follow. I got lost as I was shipping better support.

So, this commit rearranges the code - both structure and naming / comments - so that it's
clearer when we're processing the custom transform declaration vs the place where it's
used. I also call out much more clearly that, for historical reasons, we're actually
creating a synthetic decorator where none exists in the code in order to handle the
`init_subclass_args` hook that's used for base-class-style transforms.

In the process I fix a bug; the preexisting code was so unclear that we ourselves
had mixed up `kw_only` vs `kw_only_default`, which led us to file T129464224.

Reviewed By: grievejia

Differential Revision: D55977410

fbshipit-source-id: b8d1ddf65bb5624bdb27a8c537bb65f795e4e27f